requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.5
astroid==2.3.3
billiard==3.5.0.5
celery==4.0.0
colorama==0.4.1
dj-database-url==0.5.0
Django==3.0
django-autoslug==1.9.6
django-bootstrap-modal-forms==1.4.1
django-celery==3.3.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-model-utils==3.1.2
django-widget-tweaks==1.4.2
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
pandas==1.0.1
Pillow==6.2.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2017.3
selenium==3.14.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
virtualenv==16.7.7
wrapt==1.11.2
